Don't use overbroad IP range for sshuttle
Don't use the entire 10.0.0.0/8 for sshuttle as this can end up eating
traffic that has nothing to do with our cluster. Use a /20 instead.

On the bastion, we can guarantee that $NETWORK_SUBNET is set, so use
that. On the workstation, hardcode the same IP range that NETWORK_SUBNET
was set to in case the environment variables were all originally set on
the bastion.