Clarify that students want client secret value
When creating a client secret, the result includes a client secret ID
and a client secret value. Make it clearer that students should be using
the client secret value to set the CLIENT_SECRET environment variable.

In addition to wording, remove the hyphens from the sample value so that
it doesn't look like it is punctuated the way the secret ID is.