Simplify connectivity with sshutle
The original instructions were written with the assumption that the
workstation running the lab did not have access tunneled through
sshuttle, and therefore assumed podman wouldn't be able to connect to
the registry. It also instructed to ssh to the jumpbox to interact with
the registry.

Instead, just keep everything on the machine that is already being used,
with all of the environment variables already set, and let sshuttle take
care of proxying the connection to the registry.